move editions endpoints and tests into separate files
